# protocol: do not use string slicing when decoding data