Strings in Python are immutable, so using slicing here was resulting in a
massive performance penalty. This replaces the slicing with a technique
using an index that is tracked internally for the Protocol object. The
index is moved along as bytes are read from the data object.

Below is some profile data collected before and after this patch, input
was from a weechat instance with ~60 open buffers/channels. On a big
core CPU the time isn't really that meaningful, but on a weaker CPU
(e.g. ARM) the impact is tens of seconds off the start time.
This was measured on an AArch64 CPU, profiling was started as soon as
the app launched (with autoconnect enabled), and stopped as soon as the
UI was responsive (i.e. I could scroll the buffer list).

Before:
```
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
    70572   59.935    0.001   88.917    0.001 protocol.py:158(_obj_len_data)
    54632   33.741    0.001   34.297    0.001 protocol.py:185(_obj_int)
    15858    9.502    0.001    9.671    0.001 protocol.py:177(_obj_char)
        1    2.837    2.837   14.535   14.535 gtk-weechat.py:342(_parse_line)
        3    2.729    0.910  112.619   37.540 protocol.py:241(_obj_hdata)
     4153    2.445    0.001    9.688    0.002 buffer.py:102(_display_with_colors)
    24713    2.196    0.000    2.200    0.000 Gtk.py:722(create_tag)
     2467    1.862    0.001    1.882    0.001 protocol.py:149(_obj_type)
    11171    1.668    0.000    2.611    0.000 Gtk.py:762(insert_with_tags)
    13513    1.103    0.000    1.128    0.000 Gtk.py:756(insert)
    45285    1.074    0.000   57.336    0.001 protocol.py:201(_obj_str)
    72506    0.956    0.000    0.956    0.000 {method 'decode' of 'bytes' objects}
    22945    0.908    0.000   30.805    0.001 protocol.py:212(_obj_ptr)
      325    0.866    0.003    0.873    0.003 __init__.py:270(new_init)
     1820    0.777    0.000    0.863    0.000 buffer.py:190(get_text_pixel_width)
     8881    0.750    0.000    0.750    0.000 Gdk.py:89(__init__)
    95833    0.741    0.000    0.741    0.000 {built-in method _struct.unpack}
       61    0.619    0.010    1.430    0.023 buffer.py:212(__init__)
170527/170395    0.505    0.000    0.506    0.000 {built-in method builtins.len}
     2342    0.456    0.000   22.951    0.010 protocol.py:297(_obj_array)
        1    0.443    0.443  132.151  132.151 Gio.py:40(run)
     2342    0.422    0.000   11.555    0.005 buffer.py:66(display)
     6538    0.421    0.000    1.301    0.000 Gtk.py:993(_do_insert)
        1    0.366    0.366    0.498    0.498 gtk-weechat.py:61(__init__)
```

After:
```
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    2.887    2.887   14.594   14.594 gtk-weechat.py:342(_parse_line)
     4153    2.450    0.001    9.674    0.002 buffer.py:102(_display_with_colors)
    24675    2.197    0.000    2.201    0.000 Gtk.py:722(create_tag)
    11152    1.664    0.000    2.613    0.000 Gtk.py:762(insert_with_tags)
    13496    1.106    0.000    1.131    0.000 Gtk.py:756(insert)
    70568    0.959    0.000    1.628    0.000 protocol.py:158(_obj_len_data)
      325    0.849    0.003    0.856    0.003 __init__.py:270(new_init)
     1818    0.770    0.000    0.856    0.000 buffer.py:190(get_text_pixel_width)
     8862    0.736    0.000    0.736    0.000 Gdk.py:89(__init__)
        3    0.705    0.235    3.642    1.214 protocol.py:241(_obj_hdata)
       61    0.623    0.010    1.438    0.024 buffer.py:212(__init__)
    54624    0.606    0.000    0.737    0.000 protocol.py:185(_obj_int)
        1    0.447    0.447   23.192   23.192 Gio.py:40(run)
     2344    0.427    0.000   11.560    0.005 buffer.py:66(display)
     6534    0.419    0.000    1.291    0.000 Gtk.py:993(_do_insert)
        1    0.345    0.345    0.467    0.467 gtk-weechat.py:61(__init__)
       21    0.328    0.016    0.328    0.016 {built-in method builtins.compile}
    45279    0.307    0.000    1.568    0.000 protocol.py:201(_obj_str)
```